test(providers): fix azure.TestUserAgent
Test without a proxy environment variable

Authored by: @andrason
